salvando vendedores favoritos na memoria do telefone e adicionando borda verde em volta da foto de perfil

+++ lib/screens/welcomeBackScreen.dart
@@ -46,7 +46,7 @@ class _WelcomeBackScreenState extends State<WelcomeBackScreen> {
         LoggedUserInfo loggedUserInfo = LoggedUserInfo();
         Navigator.of(context).pushReplacementNamed(AppRoutes.HOMETAB);
         UserInfo userInfo = UserInfo(
-          idCliente: response.data['user']['idCliente'] as int,
+          idCliente: response.data['user']['idCliente'],
           nomeCliente: response.data['user']['nomeCliente'],
           cpfCliente: response.data['user']['cpfCliente'],
           emailCliente: response.data['user']['emailCliente'],
@@ -60,7 +60,7 @@ class _WelcomeBackScreenState extends State<WelcomeBackScreen> {
         }
         prefs.setString(SharedPreferencesConstants.loggedUserInfos, json.encode(loggedUserInfo.toJson()));
         Cliente user = Cliente(
-          idCliente: response.data['user']['idCliente'],
+          idCliente: response.data['user']['idCliente'] as int,
           nomeCliente: response.data['user']['nomeCliente'],
           cpfCliente: response.data['user']['cpfCliente'],
           emailCliente: response.data['user']['emailCliente'],
